requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
Deprecated==1.2.14
distro==1.8.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
netaddr==0.8.0
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pypng==0.20220715.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
wrapt==1.15.0
yookassa==2.4.0
